src: add GGMLSingleLayer command to insert single layer template
This commmit adds a new command which enable the selection of a single
layer in a model. This is useful for debugging one layer at a time and
inspecting/comparing the logits of the layer.

I've found this hard to remember after a while so having it as a command
makes it easier to use.

The first three lines are necessary to build the attention
key/value/index of we get the following error:
```console
/home/dbevenius/work/llama.cpp-staging/ggml/src/ggml-backend.cpp:189: GGML_ASSERT(buffer) failed
[New LWP 3262645]
[New LWP 3262641]
[Thread debugging using libthread_db enabled]
Using host libthread_db library "/lib/aarch64-linux-gnu/libthread_db.so.1".
0x0000e058f7ea7b74 in __GI___wait4 (pid=3262646, stat_loc=0x0, options=0, usage=0x0) at ../sysdeps/unix/sysv/linux/wait4.c:30
warning: 30	../sysdeps/unix/sysv/linux/wait4.c: No such file or directory
30	in ../sysdeps/unix/sysv/linux/wait4.c
196	        waitpid(child_pid, NULL, 0);
230	        ggml_print_backtrace();
189	    GGML_ASSERT(buffer);
170	    return ggml_backend_buft_is_host(ggml_backend_buffer_get_type(buffer));
1164	    GGML_ASSERT(ggml_backend_buffer_is_host(dst->buffer));
2039	    kv->set_input_k_idxs(dst, ubatch, sinfos[i_cur]);
371	    mctx->set_input_k_idxs(self_k_idxs, ubatch);
464	    inp_attn->set_input(ubatch);
508	        input->set_input(ubatch);
804	        res->set_inputs(&ubatch);
1113	        const auto * res = process_ubatch(ubatch, LLM_GRAPH_TYPE_DECODER, mctx.get(), status);
2780	    const int ret = ctx->decode(batch);
153	    if (llama_decode(ctx, batch)) {
```